Resolved codependency issues (an incredible feat)
