+++ .vscode/settings.json
@@ -57,6 +57,7 @@
   "C_Cpp_Runner.useLinkTimeOptimization": false,
   "C_Cpp_Runner.msvcSecureNoWarnings": false,
   "files.associations": {
-    "stdexcept": "cpp"
+    "stdexcept": "cpp",
+    "iostream": "cpp"
   }
 }
