+++ Source/renderables/Pressable.h
@@ -3,6 +3,8 @@
 @ DESIGNERS: Jerry Vu
 */
 
+#pragma once
+
 #include <Systems.h>
 
 class Pressable
@@ -29,35 +31,3 @@ public:
     bool isPressed();
     bool isHovering();
 };
-
-/*
-SECTION 1: CONSTRUCTORS AND DESTRUCTORS
-*/
-Pressable::Pressable()
-{
-    /*
-    NOTE:
-        - Initially set to `false` for both.
-    */
-    this->setPressed(false);
-    this->setHovering(false);
-}
-
-Pressable::~Pressable()
-{
-    /*
-    NOTE:
-        - Empty, for now.
-    */
-}
-
-/*
-SECTION 2A: SETTERS
-*/
-void Pressable::setPressed(bool pressed) { this->pressed = pressed; }
-void Pressable::setHovering(bool hovering) { this->hovering = hovering; }
-/*
-SECTION 2B: GETTERS
-*/
-bool Pressable::isPressed() { return pressed; }
-bool Pressable::isHovering() { return hovering; }
